@@ -18,13 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  flutter_svg: ^0.19.3
+  cupertino_icons: ^1.0.3
+  flutter_svg: ^0.22.0
   http: ^0.13.3
   intro_views_flutter: ^3.2.0
 
