@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 
 Widget makeText(
   String textString, {
-  Color color,
-  FontWeight fontWeight,
-  double size,
+  Color? color,
+  FontWeight? fontWeight,
+  double? size,
 }) {
   return Text(
     textString,
